Unnest the convai SDKs
Some users don't notice "javascript sdks" is expandable and there is
also React SDK. I don't see an option to force it to be expanded in
mintlify (will ask), in the meantime we can flatten the structure.

I've put react above js, and renamed "client" to "javascript" to keep it
clearer.